Implementação
Finalizando a implementação das funcionalidades

+++ Academia-com-banco-de-dados/F_GestaoDeAlunos.cs
@@ -7,6 +7,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Forms;
+using System.IO;
 
 namespace Academia_com_banco_de_dados
 {
@@ -21,13 +22,18 @@ namespace Academia_com_banco_de_dados
         {
             InitializeComponent();
         }
+        
+       
 
         private void F_GestaoDeAlunos_Load(object sender, EventArgs e)
         {
-            string vqueryDGV = @"SELECT N_IDALUNOS as 'ID',T_NOMEALUNO as 'Aluno' FROM tb_alunos";
-            dgv_selecionarAluno.DataSource = Banco.dql(vqueryDGV);
+            /* string vqueryDGV = @"SELECT N_IDALUNOS as 'ID',T_NOMEALUNO as 'Aluno' FROM tb_alunos";
+             dgv_selecionarAluno.DataSource = Banco.dql(vqueryDGV);
+            */
+            dgv_selecionarAluno.DataSource = Banco.atulizaGestaoAlunos();
             dgv_selecionarAluno.Columns[0].Width = 50;
             dgv_selecionarAluno.Columns[1].Width = 235;
+            
 
             tb_nome.Text = dgv_selecionarAluno.Rows[0].Cells[1].Value.ToString();
 
@@ -84,10 +90,29 @@ namespace Academia_com_banco_de_dados
                 tb_nome.Text, mtb_telefone.Text, cb_status.SelectedValue,cb_turma.SelectedValue, idSelecionado); ;
                 Banco.dql(vqueryAtualizaAluno);
                 MessageBox.Show("Alterações gravadas");
+                
 
 
                 
             }
+            else
+            {
+                string[] t = turma.Split(' ');
+                int vagas = Int32.Parse(t[1]);
+                if (vagas < 1)
+                {
+                    MessageBox.Show("Não ha vagas para turma selecionada, seleciona outra turma");
+                    cb_turma.Focus();
+                    return;
+                }
+                linha = dgv_selecionarAluno.Rows[0].Index;
+                string vqueryAtualizaAluno = String.Format(@"UPDATE  tb_alunos SET T_NOMEALUNO = '{0}',T_TELEFONE = '{1}',T_STATUS = '{2}'   WHERE N_IDALUNOS = {3}",
+                tb_nome.Text, mtb_telefone.Text, cb_status.SelectedValue, idSelecionado); ;
+                Banco.dql(vqueryAtualizaAluno);
+                MessageBox.Show("Alterações gravadas");
+
+            }
+            dgv_selecionarAluno.DataSource = Banco.atulizaGestaoAlunos();           
         }
 
         private void brn_fechar_Click(object sender, EventArgs e)
@@ -101,6 +126,11 @@ namespace Academia_com_banco_de_dados
             DialogResult res = MessageBox.Show("Deseja excluir ? ID", "Exclusão",MessageBoxButtons.YesNo);
             if(res == DialogResult.Yes)
             {
+               if(File.Exists(pb_foto.ImageLocation))
+                {
+                    File.Delete(pb_foto.ImageLocation);
+                }
+                
                 string vqueryDelete = string.Format(@"DELETE FROM tb_alunos WHERE N_IDALUNOS = {0}",idSelecionado);
                 Banco.dql(vqueryDelete);              
                 MessageBox.Show("Eclusão feita com sucesso");
@@ -116,8 +146,9 @@ namespace Academia_com_banco_de_dados
             {
                 // idSelecionado = dgv.Rows[dgv.SelectedRows[0].Index].Cells[0].Value.ToString();
                 // idSelecionado = dgv_selecionarAluno.SelectedRows[0].Cells[0].Value.ToString();
+                tb_nome.Text = dgv.SelectedRows[0].Cells[1].Value.ToString();
                 idSelecionado = dgv.SelectedRows[0].Cells[0].Value.ToString();
-                string vqueryCampos = String.Format(@"SELECT T_NOMEALUNO,T_TELEFONE,T_STATUS,N_IDTURMA FROM tb_alunos WHERE N_IDALUNOS = '{0}'", idSelecionado);
+                string vqueryCampos = String.Format(@"SELECT T_NOMEALUNO,T_TELEFONE,T_STATUS,N_IDTURMA,T_FOTO FROM tb_alunos WHERE N_IDALUNOS = '{0}'", idSelecionado);
 
                 DataTable dt = Banco.dql(vqueryCampos);                
                 tb_nome.Text = dt.Rows[0].Field<string>("T_NOMEALUNO");
@@ -125,6 +156,7 @@ namespace Academia_com_banco_de_dados
                 cb_status.SelectedValue = dt.Rows[0].Field<string>("T_STATUS");
                 cb_turma.SelectedValue = dt.Rows[0].Field<Int64>("N_IDTURMA");
                 turmaAutal = cb_turma.Text;
+                pb_foto.ImageLocation = dt.Rows[0].Field<string>("T_FOTO");
 
             }
         }
